move skill models to models folder
