+++ internal/registry/service/service.go
@@ -3,8 +3,8 @@ package service
 import (
 	"context"
 
+	skillmodels "github.com/agentregistry-dev/agentregistry/internal/models"
 	"github.com/agentregistry-dev/agentregistry/internal/registry/database"
-    skillmodels "github.com/agentregistry-dev/agentregistry/internal/registry/models"
 	apiv0 "github.com/modelcontextprotocol/registry/pkg/api/v0"
 )
 
@@ -23,15 +23,15 @@ type RegistryService interface {
 	// UpdateServer updates an existing server and optionally its status
 	UpdateServer(ctx context.Context, serverName, version string, req *apiv0.ServerJSON, newStatus *string) (*apiv0.ServerResponse, error)
 
-    // Skills APIs
-    // ListSkills retrieve all skills with optional filtering
-    ListSkills(ctx context.Context, filter *database.SkillFilter, cursor string, limit int) ([]*skillmodels.SkillResponse, string, error)
-    // GetSkillByName retrieve latest version of a skill by name
-    GetSkillByName(ctx context.Context, skillName string) (*skillmodels.SkillResponse, error)
-    // GetSkillByNameAndVersion retrieve specific version of a skill by name and version
-    GetSkillByNameAndVersion(ctx context.Context, skillName string, version string) (*skillmodels.SkillResponse, error)
-    // GetAllVersionsBySkillName retrieve all versions of a skill by name
-    GetAllVersionsBySkillName(ctx context.Context, skillName string) ([]*skillmodels.SkillResponse, error)
-    // CreateSkill creates a new skill version
-    CreateSkill(ctx context.Context, req *skillmodels.SkillJSON) (*skillmodels.SkillResponse, error)
+	// Skills APIs
+	// ListSkills retrieve all skills with optional filtering
+	ListSkills(ctx context.Context, filter *database.SkillFilter, cursor string, limit int) ([]*skillmodels.SkillResponse, string, error)
+	// GetSkillByName retrieve latest version of a skill by name
+	GetSkillByName(ctx context.Context, skillName string) (*skillmodels.SkillResponse, error)
+	// GetSkillByNameAndVersion retrieve specific version of a skill by name and version
+	GetSkillByNameAndVersion(ctx context.Context, skillName string, version string) (*skillmodels.SkillResponse, error)
+	// GetAllVersionsBySkillName retrieve all versions of a skill by name
+	GetAllVersionsBySkillName(ctx context.Context, skillName string) ([]*skillmodels.SkillResponse, error)
+	// CreateSkill creates a new skill version
+	CreateSkill(ctx context.Context, req *skillmodels.SkillJSON) (*skillmodels.SkillResponse, error)
 }
